Extra read from KV storage
Do an extra attempt to read the value from the store.

If we see that a single retry does not help, we should do a proper retry logic here.

Why not now? Because we do have a new implementation that is based on R2 storage and this change is simple enough.