Use simple redis cache instead of tasks over redis
Because of the changes done in #2078 (redlock) we can simplify the caching logic and remove IdempotentRunner.
Instead of caching a task on top of the lock, we move the lock to be on top of the cache.
This way we can use a simple read and write cache instead of a dealing with the status of the task.
There's a chance it has a performance impact but it's a bit hard to tell without doing a proper stress test.