Auto-fix persisted composite schema sdl
There's a possibility that the composite schema SDL contains parts of the spergraph spec. This is a problem because we want to show the public schema to the user, and the supergraph spec is not part of that. This may happen when composite schema was produced with an old version of transformSupergraphToPublicSchema or when supergraph sdl contained something new.