fix: empty "include" made all fields inaccessible
The `applyTagFilterOnSubgraphs` logic was not handling the case where the `include` field was an empty Set.
It only checked if the `include` field was `undefined` or `null`.
This caused all fields to be inaccessible.

Instead of `null | Set<string>` I refactored it to just accept `Set<string>` and check for its size.